smaller version of ASOM app (#7)
* readme

added readme with example code for running app locally

* syncing

* added readme

added readme with example code to run app locally

* include data as 'internal data'

Data used by the app was changed from being stored as 'external' data to 'internal' data. The data was removed from data/ and saved as a single .rda file in R/. This data should be automatically loaded on launch - no need to load files from data/ in app_ui.R.

* smaller version of app

The ASOM app was updated to reduce the size of the large nftbart model objects.